net: ethernet: mtk_eth_soc: widen frag_size to fix PAGE_SIZE comparison
ring->frag_size is u16 (max 65535), but on powerpc PAGE_SIZE = 65536,
making the condition frag_size <= PAGE_SIZE always true. Widen frag_size
to u32 so the branch still works correctly on 64K page architectures.

+++ drivers/net/ethernet/mediatek/mtk_eth_soc.h
@@ -963,7 +963,7 @@ struct mtk_rx_ring {
 	void *dma;
 	u8 **data;
 	dma_addr_t phys;
-	u16 frag_size;
+	u32 frag_size;
 	u16 buf_size;
 	u16 dma_size;
 	bool calc_idx_update;
